feat: update payment and rental logic with improved error handling and new features
- Refactor payment controller and service to use consistent landlordId naming
- Enhance payment service with checks for existing payments and rental status
- Update property controller to support query parameters for filtering
- Introduce pagination and sorting in property service
- Add new fields (images, latitude, longitude) to property interface and validation
- Implement rental creation and update logic with move-in dates and status checks
- Expand review functionality to include fetching all reviews and property-specific reviews
- Introduce user management enhancements including fetching current user details
- Add admin seeding script for initial setup

+++ src/modules/category/category.service.ts
@@ -1,5 +1,7 @@
 import { prisma } from "../../lib/prisma"
 import { ICategory } from "./category.interface"
+import { AppError } from "../../errors/AppError";
+import httpStatus from "http-status";
 
 const getAllCategoriesFromDB = async () => {
     const categories = await prisma.category.findMany({})
@@ -12,6 +14,11 @@ const getCategoryByIdFromDB = async (categoryId: number) => {
             id: categoryId
         }
     })
+
+    if (!categories) {
+        throw new AppError(httpStatus.NOT_FOUND, "Category not found");
+    }
+
     return categories
 }
 
@@ -20,6 +27,18 @@ const createCategoryIntoDB = async (categoryData: Omit<ICategory, 'id' | 'slug'>
 
     const categoryDescription = description ?? "";
     const slug = name.toLowerCase().replace(/\s+/g, '-');
+    const existingCategory = await prisma.category.findFirst({
+        where: {
+            OR: [
+                { name },
+                { slug },
+            ],
+        },
+    });
+
+    if (existingCategory) {
+        throw new AppError(httpStatus.CONFLICT, "Category already exists");
+    }
 
     const newCategory = await prisma.category.create({
         data: {
@@ -37,6 +56,15 @@ const udpateCategoryIntoDB = async (categoryData: Omit<ICategory, 'slug'>) => {
 
     const categoryDescription = description ?? "";
     const slug = name.toLowerCase().replace(/\s+/g, '-');
+    const category = await prisma.category.findUnique({
+        where: {
+            id
+        }
+    });
+
+    if (!category) {
+        throw new AppError(httpStatus.NOT_FOUND, "Category not found");
+    }
 
     const updatedCategory = await prisma.category.update({
         where: {
@@ -52,6 +80,16 @@ const udpateCategoryIntoDB = async (categoryData: Omit<ICategory, 'slug'>) => {
 }
 
 const deleteCategoryFromDB = async (categoryId: number) => {
+    const category = await prisma.category.findUnique({
+        where: {
+            id: categoryId
+        }
+    });
+
+    if (!category) {
+        throw new AppError(httpStatus.NOT_FOUND, "Category not found");
+    }
+
     const deletedCategory = await prisma.category.delete({
         where: {
             id: categoryId
